fix-1678

+++ Jenkinsfile
@@ -37,8 +37,11 @@ goTemplate{
         }
       }
 
-      sh('mv /home/jenkins/go/src/github.com/fabric8-services/fabric8-wit/tmp/junit.xml `pwd`')
-      junit 'junit.xml'
+    def exists = fileExists ' /home/jenkins/go/src/github.com/fabric8-services/fabric8-wit/tmp/junit.xml '
+
+    if (exists) {
+        sh('mv /home/jenkins/go/src/github.com/fabric8-services/fabric8-wit/tmp/junit.xml `pwd`')
+        junit 'junit.xml 
     }
   }
 }
